LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/events/017-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/echo_raw' failed: Received invalid WebSocket response from the server

PASS WebSockets: this, e.target, e.currentTarget, e.eventPhase
